# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Address:
    properties:
      city:
        description: City name
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      number:
        description: Building number
        type: string
      owner_id:
        description: Owner's ID
        type: integer
      owner_type:
        description: Owner's type
        type: string
      postal:
        description: Postal code
        type: string
      street:
        description: Street name
        type: string
      updatedAt:
        type: string
    type: object
...
  model.User:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/model.Address'
        description: Address is the user's address.
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Email is the user's email address.
        type: string
      id:
        type: integer
      is_admin:
        description: IsAdmin indicates whether the user is an administrator.
        type: boolean
      nachname:
        description: Nachname is the user's last name.
        type: string
      password:
        description: Password is the user's password.
        type: string
      school_id:
        description: SchoolID is the ID of the user's school.
        type: integer
      updatedAt:
        type: string
      username:
        description: Username is the user's username.
        type: string
      vorname:
        description: Vorname is the user's first name.
        type: string
    type: object
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: patrykhegenberg@gmail.com
    name: API support
    url: http://swagger.io/support
  license:
    name: Apache 2.0
    url: http://apache.com/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: Notenverwaltung API
  version: 0.0.8
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: get all users from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Fehler beim Abrufen der Users
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create a new user
      parameters:
      - description: User object that needs to be added
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Users
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create user
      tags:
      - user
  /users/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige User-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des Users
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: get one user from db by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Ungültige User-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: User nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one user by id
      tags:
      - user
    put:
      consumes:
      - application/json
      description: update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: User nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update user
      tags:
      - user
  /users/authenticate:
    get:
      consumes:
      - application/json
      description: authenticate a user with basic authentication
      parameters:
      - description: Basic Authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Erfolgreich authentifiziert
          schema:
            type: string
        "400":
          description: Ungültige Anmeldeinformationen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Ungültiges Authentifizierungsschema
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: authenticate user
      tags:
      - user
  /users/authenticate/hx:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: authenticate a user with HX credentials
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "303":
          description: See Other
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Ungültige Anmeldeinformationen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: authenticate HX user
      tags:
      - user
  /users/hx:
    post:
      consumes:
      - application/json
      description: create a new user with HX
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: School
        in: formData
        name: school
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Users
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create user with HX
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"